requirements.txt
pandas~=2.0.0
SQLAlchemy~=2.0.8
pytest~=7.2.2
pyodbc~=4.0.35
allure-pytest~=2.13.1
allure-python-commons~=2.13.1